+++ netgore/trunk/NetGore/World/Entities/IUpdateableEntity.cs
@@ -12,6 +12,6 @@ namespace NetGore
         /// </summary>
         /// <param name="imap">The map that this <see cref="IUpdateableEntity"/> is on.</param>
         /// <param name="deltaTime">Time elapsed (in milliseconds) since the last update.</param>
-        void Update(IMap imap, float deltaTime);
+        void Update(IMap imap, int deltaTime);
     }
 }
